requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
sqlalchemy
alembic
pytest
langchain
google-generativeai
python-multipart